requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
chromadb
python-dotenv
requests
beautifulsoup4
